requirements.txt
asgiref>=3.2.3
Babel>=2.8.0
certifi>=2019.11.28
chardet>=3.0.4
Click>=7.0
decorator>=4.4.1
Django>=3.0.2
django-bootstrap-datepicker-plus>=3.0.5
django-phonenumber-field>=4.0.0
django-search-views>=0.3.7
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.4.0
django_rest_passwordreset>=1.1.0
entrypoints>=0.3
flake8>=3.7.9
future>=0.18.2
geocoder>=1.38.1
idna>=2.8
mccabe>=0.6.1
psycopg2>=2.8.4
pycodestyle>=2.5.0
pyflakes>=2.1.1
PyJWT>=1.7.1
pytz>=2019.3
ratelim>=0.1.6
requests>=2.22.0
six>=1.14.0
sqlparse>=0.3.0
urllib3>=1.25.8
gunicorn>=20.0.0
whitenoise>=5.0.1
python-decouple>=3.3
dj_database_url>=0.5.0
dj-static>=0.0.6

